[feat] vote 수정 중

+++ server/underdog/src/main/java/preproject/underdog/answer/dto/answer/AnswerRespDto.java
@@ -17,7 +17,7 @@ public class AnswerRespDto {
     private Long userId;
     private String name;
     private Long questionId;
-    private Long voteCount;
+    private int voteCount;
     private LocalDateTime createdAt;
     private LocalDateTime modifiedAt;
 }
